Update: cleaning up files to comply with the norm
