+++ main.c
@@ -6,7 +6,7 @@
 /*   By: aldokezer <aldokezer@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 14:09:57 by aldokezer         #+#    #+#             */
-/*   Updated: 2024/03/21 15:40:03 by aldokezer        ###   ########.fr       */
+/*   Updated: 2024/03/21 16:14:49 by aldokezer        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 // Implement:
 
 //1) Preexec to update exit status
-//2) Preexec to work with infile and outfile (fd0 and fd01)
+//2) Preexec to work with infile and outfile (fd0 and fd01) - done
 //3) Preexec to properly clean and prepare for a new input
 
 
@@ -36,6 +36,7 @@ int	main (int argc, char *argv[], char *envp[])
 	line = NULL;
 	ft_init_cmd_tab(&cmd_tab);
 	ft_init_mini_data(&minidata, envp);
+	// Get PID to see process information
 	ft_printf("%d\n", getpid());
 	while (1)
 	{
